+++ src/mruby_integration/models/gesture.cpp
@@ -55,6 +55,21 @@ mrb_Gesture_drag_angle(mrb_state* mrb, mrb_value) -> mrb_value
   return mrb_float_value(mrb, GetGestureDragAngle());
 }
 
+auto
+mrb_Gesture_pinched(mrb_state* mrb, mrb_value) -> mrb_value
+{
+  auto* return_vector = static_cast<Vector2*>(malloc(sizeof(Vector2)));
+  *return_vector = GetGesturePinchVector();
+
+  return mrb_Vector2_value(mrb, return_vector);
+}
+
+auto
+mrb_Gesture_pinch_angle(mrb_state* mrb, mrb_value) -> mrb_value
+{
+  return mrb_float_value(mrb, GetGesturePinchAngle());
+}
+
 void
 append_models_Gesture(mrb_state* mrb)
 {
@@ -76,6 +91,13 @@ append_models_Gesture(mrb_state* mrb)
     mrb, Gesture_class, "dragged", mrb_Gesture_dragged, MRB_ARGS_NONE());
   mrb_define_class_method(
     mrb, Gesture_class, "drag_angle", mrb_Gesture_drag_angle, MRB_ARGS_NONE());
+  mrb_define_class_method(
+    mrb, Gesture_class, "pinched", mrb_Gesture_pinched, MRB_ARGS_NONE());
+  mrb_define_class_method(mrb,
+                          Gesture_class,
+                          "pinch_angle",
+                          mrb_Gesture_pinch_angle,
+                          MRB_ARGS_NONE());
 
   load_ruby_models_gesture(mrb);
 }